requirements.txt
numpy>=2.3.4
pyserial>=3.5
Pillow>=12.0.0
potracer>=0.0.4
vtracer>=0.6.11
